Forward `hash` to inner type.
This addresses issue #13.

We used to hash pointers instead of values, as for interned values hash
equality is equivalent to value equality.  This is wrong in multiple
ways.  First, this violates the requirement that `Hash` must
be equivalent in owned and borrowed values.   Second, if an interned
value is garbage collected and then created again between two `hash()`
invocations, they will return different values.